# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CitationOut": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "claim": {
            "title": "Claim",
            "type": "string"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "claim",
          "chunk_id",
          "validation_status"
        ],
        "title": "CitationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LessonContentOut": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "content_markdown": {
            "title": "Content Markdown",
            "type": "string"
          },
          "grounding_mode": {
            "title": "Grounding Mode",
            "type": "string"
          },
          "retrieved_chunk_ids": {
            "items": {
              "type": "string"
            },
            "title": "Retrieved Chunk Ids",
            "type": "array"
          }
        },
        "required": [
          "content_markdown",
          "citations",
          "grounding_mode",
          "retrieved_chunk_ids"
        ],
        "title": "LessonContentOut",
        "type": "object"
      },
...
      "TutorDone": {
        "properties": {
          "grounding_mode": {
            "title": "Grounding Mode",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "token_usage": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Token Usage",
            "type": "object"
          }
        },
        "required": [
          "message_id",
          "grounding_mode",
          "token_usage"
        ],
        "title": "TutorDone",
        "type": "object"
      },
      "TutorEvent": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorRetrieval"
              },
              {
                "$ref": "#/components/schemas/TutorToken"
              },
              {
                "$ref": "#/components/schemas/CitationOut"
              },
              {
                "$ref": "#/components/schemas/TutorDone"
              },
              {
                "$ref": "#/components/schemas/TutorInsufficient"
              }
            ],
            "title": "Data"
          },
          "event": {
            "enum": [
              "retrieval",
              "token",
              "citation",
              "done",
              "insufficient"
            ],
            "title": "Event",
            "type": "string"
          }
        },
        "required": [
          "event",
          "data"
        ],
        "title": "TutorEvent",
        "type": "object"
      },
      "TutorInsufficient": {
        "properties": {
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "text"
        ],
        "title": "TutorInsufficient",
        "type": "object"
      },
      "TutorQuestionIn": {
        "properties": {
          "context_lesson_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Lesson Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "decision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "TutorQuestionIn",
        "type": "object"
      },
      "TutorRetrieval": {
        "properties": {
          "chunk_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chunk Ids",
            "type": "array"
          }
        },
        "required": [
          "chunk_ids"
        ],
        "title": "TutorRetrieval",
        "type": "object"
      },
      "TutorToken": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TutorToken",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/lessons/{lesson_id}/content": {
      "get": {
        "description": "Real, grounded lesson content -- not a placeholder. Plain JSON, not SSE:\nunlike the tutor chat, there is no streaming UX need here, so this\nreturns TutorService's already-computed result directly.\n\n`decision_id`: pass the AdaptationDecision id when this lesson is being\nfetched because it was the recommended next activity, so the resulting\nTutorMessage carries that provenance (Phase 7).",
        "operationId": "get_lesson_content_api_v1_courses__course_id__lessons__lesson_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "detailed",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "decision_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonContentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lesson Content",
        "tags": [
          "tutor"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/tutor": {
      "post": {
        "operationId": "ask_tutor_api_v1_courses__course_id__tutor_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorQuestionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TutorEvent",
                  "type": "string"
                }
              }
            },
            "description": "SSE frames: event name plus JSON data from TutorEvent. Full answer is validated before emission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Tutor",
        "tags": [
          "tutor"
        ]
      }
    },
...
  }
}